move EventMap to event_map mod
