+++ rust/vote-tx-v2/src/gen_tx/decoding.rs
@@ -7,7 +7,7 @@ use minicbor::{
     Decode, Decoder, Encode, Encoder,
 };
 
-use super::{Choice, EventKey, EventMap, GeneralizedTx, Proof, PropId, TxBody, Uuid, VoterData};
+use super::{Choice, GeneralizedTx, Proof, PropId, TxBody, Uuid, VoterData};
 use crate::Cbor;
 
 /// UUID CBOR tag <https://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml/>.
@@ -67,74 +67,6 @@ where
             .write_all(&sign_bytes)
             .map_err(minicbor::encode::Error::write)?;
 
-        Ok(())
-    }
-}
-
-impl Decode<'_, ()> for EventMap {
-    fn decode(d: &mut Decoder<'_>, (): &mut ()) -> Result<Self, minicbor::decode::Error> {
-        let Some(len) = d.map()? else {
-            return Err(minicbor::decode::Error::message(
-                "must be a defined sized map",
-            ));
-        };
-
-        let map = (0..len)
-            .map(|_| {
-                let key = EventKey::decode(d, &mut ())?;
-
-                let value = read_cbor_bytes(d).map_err(|_| {
-                    minicbor::decode::Error::message("missing event map `value` field")
-                })?;
-                Ok((key, value))
-            })
-            .collect::<Result<_, _>>()?;
-
-        Ok(EventMap(map))
-    }
-}
-
-impl Encode<()> for EventMap {
-    fn encode<W: minicbor::encode::Write>(
-        &self, e: &mut Encoder<W>, (): &mut (),
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        e.map(self.0.len() as u64)?;
-
-        for (key, value) in &self.0 {
-            key.encode(e, &mut ())?;
-
-            e.writer_mut()
-                .write_all(value)
-                .map_err(minicbor::encode::Error::write)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl Decode<'_, ()> for EventKey {
-    fn decode(d: &mut Decoder<'_>, (): &mut ()) -> Result<Self, minicbor::decode::Error> {
-        let pos = d.position();
-        // try to decode as int
-        if let Ok(i) = d.int() {
-            Ok(EventKey::Int(i))
-        } else {
-            // try to decode as text
-            d.set_position(pos);
-            let str = d.str()?;
-            Ok(EventKey::Text(str.to_string()))
-        }
-    }
-}
-
-impl Encode<()> for EventKey {
-    fn encode<W: minicbor::encode::Write>(
-        &self, e: &mut Encoder<W>, (): &mut (),
-    ) -> Result<(), minicbor::encode::Error<W::Error>> {
-        match self {
-            EventKey::Int(i) => e.int(*i)?,
-            EventKey::Text(s) => e.str(s)?,
-        };
         Ok(())
     }
 }
@@ -313,7 +245,10 @@ mod tests {
     use proptest_derive::Arbitrary;
     use test_strategy::proptest;
 
-    use super::{super::Vote, *};
+    use super::{
+        super::{EventKey, EventMap, Vote},
+        *,
+    };
 
     type ChoiceT = Vec<u8>;
     type ProofT = Vec<u8>;
